# app/static/swagger.yaml
openapi: 3.0.0
info:
  title: Bus Payments & Rewards API
  description: Comprehensive API for bus payment system with offers, rewards, bookings, and refunds
  version: 1.0.0
servers:
  - url: http://localhost:{port}
    description: Local development server
    variables:
      port:
        default: '5006'
        description: Server port number
...
paths:
...
  /simulator/status:
    get:
      tags:
        - Integration
      summary: Get simulator status
      responses:
        '200':
          description: Simulator status information
        '500':
          description: Internal server error

  # Travel Integration
...
  /offers/travel/search-flights:
    post:
      tags:
        - Travel Integration
      summary: Search flights
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                origin:
                  type: string
                destination:
                  type: string
                departure_date:
                  type: string
                  format: date
                return_date:
                  type: string
                  format: date
                passengers:
                  type: integer
                  minimum: 1
                  default: 1
              required:
                - origin
                - destination
                - departure_date
      responses:
        '200':
          description: Available flights
  /offers/travel/book-flight:
    post:
      tags:
        - Travel Integration
      summary: Book flight
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                flight_id:
                  type: string
                customer_id:
                  type: integer
                passengers:
                  type: array
                  items:
                    type: object
              required:
                - flight_id
                - customer_id
                - passengers
      responses:
        '201':
          description: Flight booked successfully
  /offers/travel/booking/{booking_reference}:
    get:
      tags:
        - Travel Integration
      summary: Get travel booking details
      parameters:
        - in: path
          name: booking_reference
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Booking details

  # Hotel Integration
...
  /offers/hotel/search-hotels:
    post:
      tags:
        - Hotel Integration
      summary: Search hotels
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                city:
                  type: string
                check_in:
                  type: string
                  format: date
                check_out:
                  type: string
                  format: date
                guests:
                  type: integer
                  minimum: 1
                  default: 1
                rooms:
                  type: integer
                  minimum: 1
                  default: 1
              required:
                - city
                - check_in
                - check_out
      responses:
        '200':
          description: Available hotels
  /offers/hotel/book-hotel:
    post:
      tags:
        - Hotel Integration
      summary: Book hotel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                hotel_id:
                  type: string
                customer_id:
                  type: integer
                check_in:
                  type: string
                  format: date
                check_out:
                  type: string
                  format: date
                guests:
                  type: array
                  items:
                    type: object
                rooms:
                  type: integer
              required:
                - hotel_id
                - customer_id
                - check_in
                - check_out
                - guests
      responses:
        '201':
          description: Hotel booked successfully
  /offers/hotel/booking/{booking_reference}:
    get:
      tags:
        - Hotel Integration
      summary: Get hotel booking details
      parameters:
        - in: path
          name: booking_reference
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Hotel booking details

  # Shopping Integration
...
  /offers/search/travel-package:
    post:
      tags:
        - Travel Integration
      summary: Search travel packages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                origin:
                  type: string
                destination:
                  type: string
                departure_date:
                  type: string
                  format: date
                return_date:
                  type: string
                  format: date
                passengers:
                  type: integer
                  minimum: 1
                rooms:
                  type: integer
                  minimum: 1
              required:
                - origin
                - destination
                - departure_date
                - return_date
      responses:
        '200':
          description: Available travel packages